feat(db-connection):modified connection to db

+++ profile-service/src/main/java/com/digitalholics/profileservice/Profile/service/PatientServiceImpl.java
@@ -2,19 +2,25 @@ package com.digitalholics.profileservice.Profile.service;
 
 
 import com.digitalholics.profileservice.Profile.domain.model.entity.Patient;
+import com.digitalholics.profileservice.Profile.domain.model.entity.User.Role;
 import com.digitalholics.profileservice.Profile.domain.model.entity.User.User;
 import com.digitalholics.profileservice.Profile.domain.persistence.PatientRepository;
+import com.digitalholics.profileservice.Profile.domain.persistence.UserRepository;
 import com.digitalholics.profileservice.Profile.domain.service.PatientService;
 import com.digitalholics.profileservice.Profile.resource.Patient.CreatePatientResource;
 import com.digitalholics.profileservice.Profile.resource.Patient.UpdatePatientResource;
 import com.digitalholics.profileservice.Shared.Exception.ResourceNotFoundException;
 import com.digitalholics.profileservice.Shared.Exception.ResourceValidationException;
-import jakarta.validation.ConstraintViolation;
-import jakarta.validation.Validator;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
 import jakarta.ws.rs.NotFoundException;
+import org.modelmapper.internal.bytebuddy.asm.Advice;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -24,22 +30,29 @@ import org.springframework.web.client.RestTemplate;
 
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
-import java.util.Set;
 
 @Service
 public class PatientServiceImpl implements PatientService {
 
-    private static final String ENTITY = "Patient";
+;    private static final String ENTITY = "Patient";
 
     private final PatientRepository patientRepository;
+
+    private final UserRepository userRepository;
+
+    private final RestTemplate restTemplate;
 
 
     //private final Validator validator;
 
-    public PatientServiceImpl(PatientRepository patientRepository) {
+    @Autowired
+    public PatientServiceImpl(PatientRepository patientRepository, UserRepository userRepository, RestTemplate restTemplate) {
         this.patientRepository = patientRepository;
+        this.userRepository = userRepository;
         //this.validator = validator;
+        this.restTemplate = restTemplate;
     }
 
     @Override
@@ -48,8 +61,8 @@ public class PatientServiceImpl implements PatientService {
     }
 
     @Override
-    public Page<Patient> getAll(Pageable pageable) {
-        return patientRepository.findAll(pageable);
+    public Page<Patient> getAll( Pageable pageable) {
+            return patientRepository.findAll(pageable);
     }
 
     @Override
@@ -70,31 +83,17 @@ public class PatientServiceImpl implements PatientService {
     }
 
     @Override
-    public Patient getLoggedInPatient() {
-        /*Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        if (principal instanceof UserDetails) {
-            UserDetails userDetails = (UserDetails) principal;
-            String username = userDetails.getUsername();
-            return patientRepository.findPatientsByUserUsername(username);
-        }*/
-        throw new ResourceNotFoundException("No se encontró un paciente autenticado.");
+    public Patient getLoggedInPatient(String jwt) {
+
+        User user = validateJwtAndGetUser(jwt);
+
+        return patientRepository.findPatientsByUserUsername(user.getUsername());
     }
 
     @Override
-    public Patient create(CreatePatientResource patientResource) {
-       // Set<ConstraintViolation<CreatePatientResource>> violations = validator.validate(patientResource);
+    public Patient create(String jwt, CreatePatientResource patientResource) {
 
-        //if (!violations.isEmpty())
-            //throw new ResourceValidationException(ENTITY, violations);
-
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String username = authentication.getName();
-
-        //Optional<User> userOptional = userRepository.findByUsername(username);
-
-        //User user = userOptional.orElseThrow(() -> new NotFoundException("User not found for username: " + username));
-
-
+        User user = validateJwtAndGetUser(jwt);
 
         Patient patientWithDni = patientRepository.findPatientByDni(patientResource.getDni());
 
@@ -103,7 +102,7 @@ public class PatientServiceImpl implements PatientService {
                     "A patient with the same Dni first name already exists.");
 
         Patient patient = new Patient();
-        patient.setUserId(patientResource.getUserId());
+        patient.setUser(user);
         patient.setAge(patientResource.getAge());
         patient.setDni(patientResource.getDni());
         patient.setLocation(patientResource.getLocation());
@@ -116,26 +115,69 @@ public class PatientServiceImpl implements PatientService {
     }
 
     @Override
-    public Patient update(Integer patientId, UpdatePatientResource request) {
+    public Patient update(String jwt, Integer patientId, UpdatePatientResource request) {
+        User user = validateJwtAndGetUser(jwt);
+
         Patient patient = getById(patientId);
 
-        patient.setDni(request.getDni() != null ? request.getDni() : patient.getDni());
-        patient.setAge(request.getAge() != null ? request.getAge() : patient.getAge());
-        patient.setPhotoUrl(request.getPhotoUrl() != null ? request.getPhotoUrl() : patient.getPhotoUrl());
-        patient.setAppointmentQuantity(request.getAppointmentQuantity() != null ? request.getAppointmentQuantity() : patient.getAppointmentQuantity());
-        patient.setLocation(request.getLocation() != null ? request.getLocation() : patient.getLocation());
-        patient.setBirthdayDate(request.getBirthdayDate() != null ? request.getBirthdayDate() : patient.getBirthdayDate());
+        if(Objects.equals(user.getUsername(), patient.getUser().getUsername()) || Objects.equals(String.valueOf(user.getRole()), "ADMIN")){
 
-        return patientRepository.save(patient);
+            patient.setDni(request.getDni() != null ? request.getDni() : patient.getDni());
+            patient.setAge(request.getAge() != null ? request.getAge() : patient.getAge());
+            patient.setPhotoUrl(request.getPhotoUrl() != null ? request.getPhotoUrl() : patient.getPhotoUrl());
+            patient.setAppointmentQuantity(request.getAppointmentQuantity() != null ? request.getAppointmentQuantity() : patient.getAppointmentQuantity());
+            patient.setLocation(request.getLocation() != null ? request.getLocation() : patient.getLocation());
+            patient.setBirthdayDate(request.getBirthdayDate() != null ? request.getBirthdayDate() : patient.getBirthdayDate());
+
+            return patientRepository.save(patient);
+        }
+
+        throw new ResourceValidationException("JWT",
+                "Invalid access.");
     }
 
     @Override
-    public ResponseEntity<?> delete(Integer patientId) {
+    public ResponseEntity<?> delete(String jwt, Integer patientId) {
+
+        User user = validateJwtAndGetUser(jwt);
+
         return patientRepository.findById(patientId).map(patient -> {
-            patientRepository.delete(patient);
-            return ResponseEntity.ok().build();
+            if(Objects.equals(user.getUsername(), patient.getUser().getUsername()) || Objects.equals(String.valueOf(user.getRole()), "ADMIN")){
+
+                patientRepository.delete(patient);
+                return ResponseEntity.ok().build();
+            }
+            throw new ResourceValidationException("JWT",
+                    "Invalid access.");
         }).orElseThrow(()-> new ResourceNotFoundException(ENTITY,patientId));
+
     }
 
+    public User validateJwtAndGetUser(String jwt) {
+        if (jwt != null && jwt.startsWith("Bearer ")) {
+            jwt = jwt.substring(7); // Quita los primeros 7 caracteres ("Bearer ")
+        }
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.set("Authorization", jwt);
+
+        HttpEntity<String> requestEntity = new HttpEntity<>(null, headers);
+
+        String validationEndpointUrl = "http://security-service/api/v1/security/auth/validate-jwt";
+        ResponseEntity<String> responseEntity = restTemplate.exchange(validationEndpointUrl, HttpMethod.GET, requestEntity, String.class);
+
+        Optional<User> userOptional = userRepository.findByUsername(responseEntity.getBody());
+
+        User user = userOptional.orElseThrow(() -> new NotFoundException("User not found for username: " + responseEntity.getBody()));
+
+        if(Objects.equals(String.valueOf(user.getRole()), "PATIENT")
+                || Objects.equals(String.valueOf(user.getRole()), "ADMIN")){
+            return user;
+        }
+
+        throw new ResourceValidationException("JWT",
+                "Invalid rol.");
+    }
 
 }
+
